Changed: Work with db
1)Created directory for files for database
2)Changed some names
3)Deleted file sqlal and test files
4)Added file with reqirements
5)Rework of configuration
