--- config.py
@@ -1,1 +1,0 @@
-db_name = "ts_database.db"
