+++ database/README!
@@ -5,6 +5,3 @@
 В migrate.py написана функция запускающая миграцию БД,
 на вход принимае список изменений (объекты созданные с
 помощью migrate.*).
-
-Config.py содержит данные по типу названия файла БД,
-токена бота(мб в дальнейшем) и тому подобное.
