--- requirements.txt
@@ -1,13 +1,0 @@
-flake8==6.0.0
-flake8-broken-line==1.0.0
-flake8-isort==6.0.0
-flake8-plugin-utils==1.3.3
-flake8-return==1.2.0
-isort==5.12.0
-mccabe==0.7.0
-peewee==3.16.2
-pep8-naming==0.13.3
-pycodestyle==2.10.0
-pyflakes==3.0.1
-pyparsing==3.1.0
-schedule==1.2.0
